refactor(routes): restructure routes with landing page and examples
- Create new landing page with links to examples
- Move kitchen sink demo to /kitchen-sink route
- Add virtual scroll example at /virtual-scroll route
- Virtual scroll demo includes debug mode for development
